% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mle.R
\name{check_input_esc_mle}
\alias{check_input_esc_mle}
\title{Check the input of the function \code{esc_mle()}.}
\usage{
check_input_esc_mle(esc_data, approximate, assumeND)
}
\arguments{
\item{esc_data}{Dataframe. See \code{esc_mle()}.}

\item{approximate}{Logical. See \code{esc_mle()}.}

\item{assumeND}{Logical. See \code{esc_mle()}.}
}
\value{
A list containing formated \code{concentrations} and \code{cqs}.
}
\description{
Check the input of the function \code{esc_mle()}.
}
