% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{checks_mcmc}
\alias{checks_mcmc}
\title{Check parameters for \code{conc_mcmc()}}
\usage{
checks_mcmc(cqs, model, level)
}
\arguments{
\item{cqs}{Numeric vector of Cq values from sample replicates, non-detects
coded as NaN}

\item{model}{esc object representing fitted model to use for estimation}

\item{level}{Desired credible level.}
}
\value{
Nothing.
}
\description{
Check parameters for \code{conc_mcmc()}
}
