% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{checks_mcmc_multi}
\alias{checks_mcmc_multi}
\title{Check parameters for \code{multi_conc_mcmc()}}
\usage{
checks_mcmc_multi(cq_data, model, level)
}
\arguments{
\item{cq_data}{data frame with an sample column specifying the names of the
samples from which reactions were generated, and a cqs column containing the
corresponding Cq values. Cq values must be numeric, with non0detects encoded
as NaN.}

\item{model}{esc object representing fitted model to use for estimation}

\item{level}{Desired credible level.}
}
\value{
Nothing.
}
\description{
Check parameters for \code{multi_conc_mcmc()}
}
