% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{conc_interval}
\alias{conc_interval}
\title{Generate credible interval for concentration analytically}
\usage{
conc_interval(cqs, model, level = 0.95)
}
\arguments{
\item{cqs}{Numeric vector of Cq values from sample replicates, non-detects
coded as NaN}

\item{model}{esc object representing fitted model to use for estimation}

\item{level}{Desired credibily level, defaults to 0.95}
}
\value{
List specifying the lower and upper bounds of desired credible
interval, as well as the maximum likelihood concentration estimate
}
\description{
Given list of Cq values from a set of technical replicates and fitted ESC
model, generates credible interval of desired level through numerical
integration of the posterior distribution
}
