% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{conc_likelihood_factory}
\alias{conc_likelihood_factory}
\title{Factory Function to Create Concentration Likelihood Function}
\usage{
conc_likelihood_factory(cqs, intercept, slope, sigma)
}
\arguments{
\item{cqs}{Numeric vector of observed Cq values, with non-detects coded as NaN}

\item{intercept}{intercept parameter of ESC model}

\item{slope}{slope parameter of ESC model}

\item{sigma}{sigma parameter of ESC model}
}
\value{
Function that takes single input concentration and returns
corresponding likelihhod
}
\description{
Given ESC Model parameters and observed Cq data, returns function implementing
likelihood calculation for concentration estimation
}
