% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{conc_mcmc}
\alias{conc_mcmc}
\title{MCMC Estimation of Concentrations Using ESC Model}
\usage{
conc_mcmc(cqs, model, level = 0.95)
}
\arguments{
\item{cqs}{Numeric vector of Cq values from sample replicates, non-detects
coded as NaN}

\item{model}{esc object representing fitted model to use for estimation}

\item{level}{Desired credible level, defaults to 0.95}
}
\value{
A list of two elements. The first, named interval, is a list
containing the bounds of the credible interval, along with the mean and median
of the generated posterior samples. The second, named mcmc_samples, is the
object produced by the runjags package representing the generated samples
}
\description{
Given list of Cq values from a set of technical replicates and fitted ESC
model, generates credible interval of desired level by approximating the
posterior distribution for concentration using MCMC sampling. Requires
installation of JAGS and the rjags and runjags packages
}
