% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{conc_mle}
\alias{conc_mle}
\title{Estimate concentration from Cq values}
\usage{
conc_mle(cqs, model)
}
\arguments{
\item{cqs}{Numeric vector of Cq values from sample replicates, non-detects
coded as NaN}

\item{model}{\code{esc} object representing fitted model to use for estimation.}
}
\value{
MLE of concentration
}
\description{
Given list of Cq values from a set of technical replicates and fitted ESC
model, generates maximum likelihood estimate of concentration
}
