% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esc.R
\name{cq_quantile_est}
\alias{cq_quantile_est}
\title{Faster Estimation of Theoretical Quantiles for Cq Values}
\usage{
cq_quantile_est(alpha, conc, intercept, slope, sigma)
}
\arguments{
\item{alpha}{quantile level}

\item{intercept}{intercept parameter of ESC model}

\item{slope}{slope parameter of ESC model}

\item{sigma}{sigma parameter of ESC model}

\item{concentrations}{numeric concentration at which to evaluate specified
quantile}
}
\value{
estimate of specified quantile
}
\description{
Given ESC model parameters, quantile level alpha, and concentration conc,
estimates that alpha-th quantile of the theoretical distribution of Cq values
at the specified concentration. For efficiency, ensures that a maximum of 100
N0 values are considered when the computation is performed. N0 values used for
computation are equally spaced out between the computed lower and upper bounds,
and are used to estimate contribution of nearby N0 values not included in the
computation.
}
