% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_class.R
\name{gen_esc}
\alias{gen_esc}
\title{Create ESC Objects With Specified Parameters}
\usage{
gen_esc(intercept, eff, sigma)
}
\arguments{
\item{intercept}{Numeric. Intercept parameter of ESC model.}

\item{eff}{Numeric. qPCR efficiency.}

\item{sigma}{Numeric. Sigma parameter of ESC model.}
}
\value{
\code{esc} object
}
\description{
Creates \code{esc} object, given parameters intercept, eff, and sigma
}
\examples{

m = gen_esc(intercept = 39, eff = 0.98, sigma = 0.2)
print_esc_model(m)

}
