% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mle.R
\name{log_likelihood_est}
\alias{log_likelihood_est}
\title{Faster Estimation of log likelihood values for ESC model}
\usage{
log_likelihood_est(conc, cqs, intercept, slope, sigma)
}
\arguments{
\item{conc}{numeric concentration}

\item{cqs}{numeric vector containing observed Cq values to estimate the
likelihood of, with non-detects encoded as NaN}

\item{intercept}{intercept parameter of ESC model}

\item{slope}{slope parameter of ESC model}

\item{sigma}{sigma parameter of ESC model}
}
\value{
estimated log likelihood values
}
\description{
Estimates likelihood of given set of observed Cq values given a concentration
and ESC model parameters. For efficiency, ensures that a maximum of 100 N0
values are considered when the computation is performed. N0 values used for
computation are equally spaced out between the computed lower and upper bounds,
and are used to estimate contribution of nearby N0 values not included in the
computation.
}
