% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{multi_conc_mcmc}
\alias{multi_conc_mcmc}
\title{MCMC Estimation of Concentrations for Multiple Samples at Once}
\usage{
multi_conc_mcmc(cq_data, model, level = 0.95)
}
\arguments{
\item{cq_data}{data frame with an sample column specifying the names of the
samples from which reactions were generated, and a cqs column containing the
corresponding Cq values. Cq values must be numeric, with non0detects encoded
as NaN.}

\item{model}{esc object representing fitted model to use for estimation}

\item{level}{Desired credible level, defaults to 0.95}
}
\value{
A list of two elements. The first, named interval, is a data frame
containing the bounds of the credible intervals, along with the means and
medians of the generated posteriorsamples. The second, named mcmc_samples,
is the object produced by the runjags package representing the generated
samples.
}
\description{
Computes Baysian credible intervals for the concentrations of multiple samples
at once, given Cq data for each sample and a single esc model object and
specified credible level to use for all the intervals. Uses MCMC sampling to
approximate the posterior distribution. Requires installation of JAGS and the
rjags and runjags packages
}
