% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_int.R
\name{new_conc_int}
\alias{new_conc_int}
\title{Generate New conc_int Object}
\usage{
new_conc_int(mle, interval, grid, pdf, cdf, level)
}
\arguments{
\item{mle}{MLE estimate of concentration}

\item{interval}{Numeric vector representing credible interval: first element
is the lower endpoint, second is the upper endpoint}

\item{grid}{Numeric vector containing grid of value at with PDF/CDF were
evaluated}

\item{pdf}{Numeric vector of PDF values evaluated at each point in grid}

\item{cdf}{Numeric vector of CDF values evaluated at each point in grid}

\item{level}{Numeric. Level of the credible interval}
}
\value{
Object of class conc_int
}
\description{
Creates new object of class conc_int, containing a list specifying the upper
and lower endpoints of the credible interval,the MLE concentration
estimate, and the level of the interval, as well as a data frame containing
the PDF and CDF of posterior concentration distribution evaluated at a grid
of points
}
