% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conc_int.R
\name{plot_conc_int}
\alias{plot_conc_int}
\title{Plot Concentration Credible Intervals}
\usage{
plot_conc_int(interval, type)
}
\arguments{
\item{interval}{conc_int object containing concentration posterior distribution
and credible interval}

\item{type}{distribution function to plot, either "pdf" or "cdf"}
}
\value{
ggplot object representing generated plot
}
\description{
Given conc_int object, plots either posterior pdf or cdf for concentration,
along with MLE estimate and credible interval encoded in conc_int object
}
