% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esc.R
\name{plot_esc_data}
\alias{plot_esc_data}
\title{Plot Standard Curve Calibration Data}
\usage{
plot_esc_data(esc_data)
}
\arguments{
\item{esc_data}{Data frame containing data to bt used to fit the model. Must
contain a column named \code{concentrations} with known sample concentrations,
and a column named \code{cqs} with corresponding Cq values.
Non-detects should be encoded by a Cq value of \code{NaN}.}
}
\value{
A \code{ggplot} object representing plot.
}
\description{
Given a data frame containing concentrations
and corresponding Cq values, plots Cq values against
log-transformed concentrations.
}
