% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esc.R
\name{plot_esc_model}
\alias{plot_esc_model}
\title{Plot Fitted ESC Model}
\usage{
plot_esc_model(model, PI = 0.95, approximate = TRUE)
}
\arguments{
\item{model}{\code{esc} object representing fitted model.}

\item{PI}{Numeric. Width of the probability interval (must be between 0 and 1).}

\item{approximate}{Logical. If \code{TRUE} (the default), a faster but potentially
less accurate approximation for the likelihood function will be used at high
concentrations.}
}
\value{
A \code{ggplot} object representing plot.
}
\description{
Plots the fitted median and probability interval of
an ESC model Cq vs concentration.
}
