% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_esc_model.R
\name{print_esc_model}
\alias{print_esc_model}
\title{Print the parameters of an ESC model.}
\usage{
print_esc_model(model)
}
\arguments{
\item{show.data}{}
}
\value{
Print the model parameters.
}
\description{
Print the parameters of an ESC model.
}
