% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_cqs}
\alias{sim_cqs}
\title{Simulate Cq values based on ESC Model}
\usage{
sim_cqs(concentrations, eff, cq1, sigma)
}
\arguments{
\item{concentrations}{Vector of input concentrations}

\item{eff}{PCR Efficiency}

\item{cq1}{Cq value for initial copy number of 1}

\item{sigma}{Standard Deviation of Gaussian error in Cq values}
}
\value{
Numeric vector of Cq Values
}
\description{
Generate simulated Cq values given vector of concentrations and parameters
for the ESC Model. Non-detects are coded as NaN
}
