% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_theta.R
\name{theta_from_slope}
\alias{theta_from_slope}
\title{Calculate theta using the slope .}
\usage{
theta_from_slope(slope)
}
\arguments{
\item{slope}{Numeric.}
}
\value{
The theta parameter.
}
\description{
This is an internal function to conveniently parametrize
the slope such that the implied efficiency is naturally
constrained between 0 and 1.
\code{Efficiency = 1 / (1 + exp(-theta))} and
\code{slope = 1 / log10(1+Efficiency)}
}
