% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_est.R
\name{conc_interval}
\alias{conc_interval}
\title{Generate credible interval for concentration analytically}
\usage{
conc_interval(cqs, model, level = 0.95, approximate = TRUE)
}
\arguments{
\item{cqs}{Numeric vector of Cq values from sample replicates, non-detects
coded as \code{NaN}.}

\item{model}{A \code{esc} object representing a fitted model to use for estimation.}

\item{level}{Desired credibility level, defaults to 0.95.}

\item{approximate}{Logical. If \code{TRUE} (the default), a faster but potentially
less accurate approximation for the likelihood function will be used at high
concentrations.}
}
\value{
A \code{conc_int} object, containing a list interval specifying the lower and
upper bounds of desired credible interval, as well as the maximum likelihood
concentration estimate, and a data frame distribution, containing values of
the probability density function  and the cumulative density function
of the posterior distribution evaluated at specified points.
}
\description{
Given list of Cq values from a set of technical replicates
and a fitted ESC model, generates credible interval of desired level
through numerical integration of the posterior distribution.
}
\examples{

 esc_data = data.frame(
     concentrations = c(1,1,10, 10, 100, 500, 500),
     cqs = c(40.2, 39.3, 35.9, 36.4, 32.6, 30.0, 31.1))
 mod = esc_mle(esc_data)

 new.cqs = c( 35, NaN, 36)
 x = conc_interval(cqs = new.cqs, model = mod)
 print(x$interval)

}
