% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esc.R
\name{cq_quantile}
\alias{cq_quantile}
\title{Calculate Theoretical Quantiles for Cq Values}
\usage{
cq_quantile(concentrations, alpha, intercept, slope, sigma)
}
\arguments{
\item{concentrations}{numeric vector of concentrations at which to evaluate
specified quantile}

\item{alpha}{quantile level}

\item{intercept}{intercept parameter of ESC model}

\item{slope}{slope parameter of ESC model}

\item{sigma}{sigma parameter of ESC model}
}
\value{
Numeric vector containing the calculated quantiles
}
\description{
Given ESC model parameters, quantile level alpha, and vector of concentrations
evaluates that alpha-th quantile of the theoretical distribution of cq values
at each of the specified concentrations
}
