% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mle.R
\name{esc_log_likelihood}
\alias{esc_log_likelihood}
\title{Log Likelihood function for ESC Modle fitting}
\usage{
esc_log_likelihood(params, concentrations, cqs)
}
\arguments{
\item{params}{numeric vector containing parameters of the model: first
element is intercept, second is slope and third is sigma}

\item{concentrations}{numeric vector of concentrations that model is being
fit to}

\item{cqs}{numeric vector containing Cq value for each specified concentration,
with non-detects coded as NaN}
}
\value{
negative log likelihood evalueated at the given parameters
}
\description{
Log Likelihood function for ESC Modle fitting
}
