% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mle.R
\name{esc_mle}
\alias{esc_mle}
\title{Fit ESC Model Using MLE}
\usage{
esc_mle(concentrations, cqs)
}
\arguments{
\item{concentrations}{Numeric vector of known sample concentrations}

\item{cqs}{Numeric vector of corresponding Cq values, with non-detects coded
as NaN}
}
\value{
esc object representing fitted model
}
\description{
Fit ESC Model Using MLE
}
