% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mle.R
\name{esc_mle}
\alias{esc_mle}
\title{Fit ESC Model Using Maximum Likelihood Estimation.}
\usage{
esc_mle(
  esc_data,
  approximate = TRUE,
  assumeND = TRUE,
  nlm.print.level = 0,
  nlm.iterlim = 200
)
}
\arguments{
\item{esc_data}{Data frame containing data to be used to fit the model. Must
contain a column named \code{concentrations} with known sample concentrations, and
a column named \code{cqs} with corresponding Cq values. Non-detects should be
encoded by a Cq value of \code{NaN}.}

\item{approximate}{Logical. If \code{TRUE} (the default), a faster but potentially
less accurate approximation for the likelihood function will be used at high
concentrations.}

\item{assumeND}{Logical. If \code{TRUE} (the default), any non numerical
Cq values in the dataframe \code{esc_data} is assumed a non-detect (ND).}

\item{nlm.print.level}{Integer. Display intermediary steps of
the \code{nlm()} minimization (0, the default, silences print out).}

\item{nlm.iterlim}{Integer. Maximum number of iterations
for \code{nlm()} minimization.}
}
\value{
An \code{esc} object representing fitted model.
}
\description{
Fit the parameters of an ESC model to qPCR standard curve data. The ESC model is
described in
\href{https://doi.org/10.3389/fmicb.2023.1048661}{Schmidt et al. 2022}.
}
\examples{
esc_data = data.frame(
     concentrations = c(1,1,10, 10, 100, 500, 500),
     cqs = c(40.2, 39.3, 35.9, 36.4, 32.6, 30.0, 31.1))
mod = esc_mle(esc_data)

}
