% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esc_mle.R
\name{esc_mle}
\alias{esc_mle}
\title{Fit ESC Model Using Maximum Likelihood Estimation.}
\usage{
esc_mle(esc_data, approximate = TRUE)
}
\arguments{
\item{esc_data}{Data frame containing data to be used to fit the model. Must
contain a column named \code{concentrations} with known sample concentrations, and
a column named \code{cqs} with corresponding Cq values. Non-detects should be
encoded by a Cq value of \code{NaN}.}

\item{approximate}{Logical. If \code{TRUE} (the default), a faster but potentially
less accurate approximation for the likelihood function will be used at high
concentrations.}
}
\value{
An \code{esc} object representing fitted model.
}
\description{
Fit the parameters of an ESC model to qPCR standard curve data. The ESC model is
described in
\href{https://doi.org/10.3389/fmicb.2023.1048661}{Schmidt et al. 2022}.
}
