% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_theta.R
\name{mylogistic}
\alias{mylogistic}
\title{Logistic function.}
\usage{
mylogistic(x, minvalue, maxvalue, shallowness = 25)
}
\arguments{
\item{x}{Numeric}

\item{minvalue}{Minimum value.}

\item{maxvalue}{Maximum value.}

\item{shallowness}{Numeric. Determines how steep the logistic function is.}
}
\value{
logistic(x)
}
\description{
Logistic function.
}
