% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_theta.R
\name{mylogistic_inverse}
\alias{mylogistic_inverse}
\title{Inverse of the logistic function.}
\usage{
mylogistic_inverse(y, minvalue, maxvalue)
}
\arguments{
\item{y}{Numeric}

\item{minvalue}{Minimum value.}

\item{maxvalue}{Maximum value.}
}
\value{
logistic^-1(y)
}
\description{
Inverse of the logistic function.
}
