% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conc_post.R
\name{plot_conc_post}
\alias{plot_conc_post}
\title{Plot Credible Intervals Concentration}
\usage{
plot_conc_post(
  interval,
  type,
  title = paste("Concentration Posterior", toupper(type))
)
}
\arguments{
\item{interval}{a \code{conc_int} object containing the concentration posterior distribution
and credible interval.}

\item{type}{distribution function to plot, either \code{"pdf"} or \code{"cdf"}.}

\item{title}{String. Title for the plot. If unspecified, defaults to "Concentration
Posterior PDF" or "Concentration Posterior CDF", depending on the value of \code{type}}
}
\value{
A \code{ggplot} object representing generated plot.
}
\description{
Given a \code{conc_int} object, plots either the posterior
probability density function or the cumulative density function
for the concentration,
along with the MLE estimate and the credible interval encoded
in the \code{conc_int} object
}
