% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esc.R
\name{plot_esc_data}
\alias{plot_esc_data}
\title{Plot Standard Curve Calibration Data}
\usage{
plot_esc_data(concentrations, cqs)
}
\arguments{
\item{concentrations}{Numeric vector of known sample concentrations}

\item{cqs}{Numeric vector of corresponding Cq values, with non-detects coded
as NaN}
}
\value{
ggplot object representing plot
}
\description{
Given vectors of concentrations and corresponding Cq values, plots Cq values
agains log-transformed concentrations
}
