% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_esc.R
\name{plot_esc_model}
\alias{plot_esc_model}
\title{Plot Fitted ESC Model}
\usage{
plot_esc_model(model, plot, level = 0.95)
}
\arguments{
\item{model}{esc object represeting fitted model}

\item{plot}{ggplot object to which fitted model will be added. Optional,
if none is provided, a plot of the data used to fit the model will be
automatically generated using \code{\link[=plot_esc_data]{plot_esc_data()}}.}

\item{level}{level of probability interval, defaults to 0.95}
}
\value{
ggplot object representing plot
}
\description{
Adds fitted median and probability interval of ESC model to plot of Cq vs
Concentration
}
