% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_esc_model.R
\name{print_esc_model}
\alias{print_esc_model}
\title{Print the parameters of an ESC model.}
\usage{
print_esc_model(model)
}
\arguments{
\item{model}{A model of class \code{esc}.}
}
\value{
Print the model parameters.
}
\description{
Print the parameters of an ESC model.
}
\examples{
esc_data = data.frame(
     concentrations = c(1,1,10, 10, 100, 500, 500),
     cqs = c(40.2, 39.3, 35.9, 36.4, 32.6, 30.0, 31.1))
mod = esc_mle(esc_data)
print_esc_model(mod)

}
